Revert checking for `User::session_id`
The `session_id` is set once the user is connected to the chat room. This however requires a logged in user. If a user is inactive, it can't login. One could possibly add a field `User::active: bool` to add this functionality. See [this answer](https://stackoverflow.com/a/19533025/3131852) from the author of Flask for more info.